requirements.txt
aniso8601==8.0.0
attrs>=22.2.0
blinker==1.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
flask-restx==1.3.0
Flask-Testing==0.8.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==6.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
MarkupSafe==1.1.1
pkgutil_resolve_name==1.3.10
pyrsistent==0.16.0
python-dotenv==1.0.0
pytz==2020.1
referencing==0.32.1
requests==2.24.0
rpds-py==0.16.2
six==1.15.0
urllib3==1.25.9
Werkzeug==0.16.1
zipp==3.1.0
